+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - `Disruptor` constructors using Executor have been removed. Use ThreadFactory instead.
     - FatalExceptionHandler and IgnoreExceptionHandler now use the JDK 9 Platform Logging API, i.e. System.Logger
     - Add rewind batch feature to the BatchEventProcessor
+    - Added documentation to EventPoller
 
 ## 3.4.3